BACKEND/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.2.0
django-filter==23.3
djangorestframework-simplejwt==5.3.0
Pillow>=10.4.0
python-decouple==3.8

django-storages==1.14.2
boto3==1.34.0
gunicorn==21.2.0
whitenoise==6.6.0
django-cleanup==8.0.0
django-ckeditor==6.7.0
django-countries==7.5.1
celery>=5.3.6
redis>=5.0.1